src/services/vector-db/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
python-dotenv==1.0.0
sentence-transformers==2.3.1
numpy==1.26.3
faiss-cpu==1.7.4
pydantic==2.5.3
